+++ src/main/java/br/com/lmello/secret_santa/dto/DrawEmailDTO.java
@@ -2,10 +2,12 @@ package br.com.lmello.secret_santa.dto;
 
 import br.com.lmello.secret_santa.model.Participant;
 
+import java.math.BigDecimal;
+
 public record DrawEmailDTO(
         Participant sender,
         Participant receiver,
         String drawCode,
-        int budget
+        BigDecimal budget
 ) {
 }
